Split spawning module into subfiles
