+++ src/spawning/objects.rs
@@ -1,0 +1,12 @@
+pub mod doorway;
+pub mod grass;
+pub mod npc;
+pub mod player;
+pub mod point_light;
+pub mod primitives;
+pub mod roof;
+pub mod roof_left;
+pub mod roof_right;
+pub mod sunlight;
+pub mod util;
+pub mod wall;
