+++ src/spawning/objects/doorway.rs
