+++ src/spawning/objects/npc.rs
