+++ src/spawning/objects/point_light.rs
