+++ src/spawning/objects/roof.rs
