+++ src/spawning/objects/roof_left.rs
