+++ src/spawning/objects/roof_right.rs
