+++ src/spawning/objects/sunlight.rs
