+++ src/spawning/objects/util.rs
