+++ src/spawning/objects/wall.rs
